third_party/WebKit/LayoutTests/http/tests/devtools/elements/edit/insert-node-collapsed.html

<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
<html>
<head>
<script src="../../../inspector/inspector-test.js"></script>
<script src="../../../inspector/elements-test.js"></script>
<script>

function appendChild()
{
    var container = document.getElementById("container");
    var child = document.createElement("div");
    child.setAttribute("id", "appended");
    container.appendChild(child);
}

function test() {
  var containerNode;

  TestRunner.runTestSuite([
    function testDumpInitial(next) {
      function callback(node) {
        containerNode = node;
        TestRunner.addResult('========= Original ========');
        ElementsTestRunner.dumpElementsTree(containerNode);
        next();
      }
      ElementsTestRunner.selectNodeWithId('container', callback);
    },

    function testAppend(next) {
      function callback() {
        ElementsTestRunner.firstElementsTreeOutline().runPendingUpdates();
        TestRunner.deprecatedRunAfterPendingDispatches(function() {
          TestRunner.addResult('======== Appended =========');
          ElementsTestRunner.dumpElementsTree(containerNode);
          next();
        });
      }
      TestRunner.evaluateInPage('appendChild()', callback);
    }
  ]);
}

</script>
</head>

<body onload="runTest()">
<p>
Tests that elements panel updates hasChildren flag upon adding children to collapsed nodes.
</p>

<div id="container"></div>

</body>
</html>
