third_party/WebKit/LayoutTests/http/tests/devtools/elements/edit/insert-node-collapsed-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that elements panel updates hasChildren flag upon adding children to collapsed nodes.


Running: testDumpInitial
========= Original ========
  <div id="container"></div>

Running: testAppend
======== Appended =========
+ <div id="container">…</div>
